requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.9.12
django-mptt==0.8.7
djangorestframework==3.5.4
gunicorn==19.7.1
olefile==0.44
packaging==16.8
pillow==4.1.0
pyparsing==2.2.0
six==1.10.0
Unidecode==0.4.20
wheel==0.24.0
psycopg2==2.6.2
